+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 * added __len__ to FutureCollection
 * fixed a bug in DictionaryEQAble
 * fixed a bug in ListDeleter
+* minor breaking change: changed semantics of ListDeleter